# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bomish API",
    "version": "1.0.0",
    "description": "Same-origin cookie sessions. Mutations require Origin and X-CSRF-Token. All money in rials and weight in grams. Development providers only."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/session": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/request": {
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "devCode": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "resendAfter": {
                      "type": "integer",
                      "format": "int64"
                    }
                  },
                  "required": [
                    "ok",
                    "devCode",
                    "expiresIn",
                    "resendAfter"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                },
                "required": [
                  "phone"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "phone",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/staff/login": {
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/products": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/articles": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/addresses": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Address"
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/products/{slug}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/articles/{slug}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/cart": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/cart/items/{id}": {
      "put": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "packageId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "required": [
                  "packageId",
                  "quantity"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/checkout/quote": {
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "province": {
                    "type": "string"
                  }
                },
                "required": [
                  "province"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/checkout": {
      "post": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "orderId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Not allowed"
          },
          "409": {
            "description": "State changed"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "$ref": "#/components/schemas/Address"
                  },
                  "expectedTotalRials": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "idempotencyKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "address",
                  "expectedTotalRials",
                  "idempotencyKey"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "additionalProperties": false
      },
...
      "Cart": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "subtotalRials": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "items",
          "subtotalRials"
        ],
        "additionalProperties": false
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "csrf": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "development": {
            "type": "boolean"
          }
        },
        "required": [
          "csrf",
          "authenticated",
          "role",
          "development"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "OK": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bomish_session"
      }
    }
  }
}